modyfied commitment table and receivable table

+++ models.py
@@ -34,14 +34,14 @@ class Friend(db.Model):
 
 class Commitment(db.Model):
     id = db.Column(db.Integer, primary_key=True)
-    amount = db.Column(db.Integer, nullable=False)
+    amount = db.Column(db.Float, nullable=False)
     title = db.Column(db.String(100), nullable=False)
     date = db.Column(db.DateTime)
     friend_id = db.Column(db.Integer, db.ForeignKey('friend.id'), nullable=False)
 
 class Receivable(db.Model):
     id = db.Column(db.Integer, primary_key=True)
-    amount = db.Column(db.Integer, nullable=False)
+    amount = db.Column(db.Float, nullable=False)
     title = db.Column(db.String(100), nullable=False)
     date = db.Column(db.DateTime)
     friend_id = db.Column(db.Integer, db.ForeignKey('friend.id'), nullable=False)
